refa: moves the contributors controller into organizations module
